Login System x

+++ src/App.jsx
@@ -1,15 +1,19 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Menu from "./Pages/Menu";
 import Checkout from "./Pages/co";
-import { useState } from "react";
+import { use, useState } from "react";
 import product from "./data/produck";
 import Deskripsi from "./Pages/deskripsi";
 import Login from "./Pages/login";
+import user from "./user/user";
+import Admin from "./admin/dashboard";
+import Kasir from "./kasir/dashboard";
 
 
 export default function App() {
     const [cart, setCart] = useState([]);
     const [category, setCategory] = useState("all");
+    const [role, setRole] = useState("guest")
 
     const filtered = 
     category === "all"
@@ -55,11 +59,14 @@ const totalHarga = cart.reduce((total, item) => {
 return (
     <BrowserRouter>
   <Routes>
+    
     <Route path="/" element={<Menu cart={cart} setCart={setCart} deleteItem={deleteItem} incItem={incItem} totalHarga={totalHarga} filtered={filtered} setCategory={setCategory} />} />
+    <Route path="/admin" element={<Admin deleteItem={deleteItem} incItem={incItem}/>} />
+    <Route path="/kasir" element={<Kasir deleteItem={deleteItem} incItem={incItem}/>} />
     <Route path="/deskripsi" element={<Deskripsi setCart={setCart} cart={cart} totalHarga={totalHarga} total={filtered} incItem={incItem} deleteItem={deleteItem} />} />
     <Route path="/product/:id" element={<Deskripsi cart={cart} setCart={setCart} incItem={incItem} deleteItem={deleteItem} totalHarga={totalHarga}/>}/>
     <Route path="/checkout" element={<Checkout cart={cart} setCart={setCart} totalHarga={totalHarga} deleteItem={deleteItem} incItem={incItem} filtered={filtered}/>} />
-    <Route path="/login" element={<Login cart={cart} setCart={setCart} incItem={incItem} deleteItem={deleteItem} totalHarga={totalHarga}/> } />
+    <Route path="/login" element={<Login role={role}  setRole={setRole} cart={cart} setCart={setCart} incItem={incItem} deleteItem={deleteItem} totalHarga={totalHarga}/> } />
   </Routes>
   </BrowserRouter>
   )
